SVD_Tool_Kit/__init__.py
import numpy
import scipy
import quaternion
from .T_SVD import T_SVD,t_product,transpose_Tensor
from .Q_SVD import Q_SVD, frob_for_quaternions,dot_product_quat
from .SVD import svd
from .Compact_SVD import compact_svd
from .HO_SVD import high_Order_SVD
from .GSVD import GSVD
from .Join_SVD import join_SVD